moved the surveylist and instrumentlist to their own folders
