+++ package.json
@@ -31,7 +31,6 @@
     "lint-fix"
   ],
   "dependencies": {
-    "@google-cloud/logging": "^11.3.0",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
     "@types/express": "^5.0.6",
