+++ src/client/pages/InstrumentListPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./InstrumentList";
